refactor: 사용자 정보 수정 로직 분리
